Made changes to how seed/requests work
Full description:
The client will now send a status message that contains a list.
Each item in the list will be a tuple of the for (infoHash,status).
status = 1 if seeding that particular page/file
status = 0 if incomplete and client wants to request for that infoHash
status = 2 if incomplete and client doesnt want to request for that infoHAs